split inputgen
